fix: split detectImportPackageFromDir into config-rendering-node.ts to avoid fs in browser bundles
config-rendering.ts had top-level fs/path imports that caused Turbopack to fail
when bundling for browser contexts (dashboard github-config-push). Moved the
Node-only detectImportPackageFromDir to a separate config-rendering-node.ts file
and updated all server-side callers to import from the new path.
